fix(documents): stop paginating at limit
The documents search function does not appear to respect the limit
parameter passed to it. While it does correctly pass it to the API, it
also keeps paginating on the returned cursor until the API says stop, so
it tends to return 1000 items regardless of the value of the limit
parameter.

This commit changes the pagination loop to stop when we have received
`limit` or more result items.

See slack thread:
https://cognitedata.slack.com/archives/CKY04V4CA/p1757496077990499